+++ docs/options/options.js
@@ -6,7 +6,7 @@
   };
 
   // <define:process.env>
-  var define_process_env_default = { BUILD_TIME: "2023-03-13T15:44:40.588Z", VERSION: "0.3.7", PROD: "1", REDIRECT_URL: "https://immersive-translate.owenyoung.com/auth-done/", IMMERSIVE_TRANSLATE_INJECTED_CSS: `:root {
+  var define_process_env_default = { BUILD_TIME: "2023-03-14T21:00:13.869Z", VERSION: "0.3.8", PROD: "1", REDIRECT_URL: "https://immersive-translate.owenyoung.com/auth-done/", IMMERSIVE_TRANSLATE_INJECTED_CSS: `:root {
   --immersive-translate-theme-underline-borderColor: #72ece9;
   --immersive-translate-theme-nativeUnderline-borderColor: #72ece9;
   --immersive-translate-theme-nativeDashed-borderColor: #72ece9;
@@ -303,7 +303,7 @@
   display: inline-block !important;
   margin: 0 4px !important;
   border: 2px rgba(221, 244, 255, 0.6) solid !important;
-  border-top: 2px rgba(9, 105, 218, 0.8) solid !important;
+  border-top: 2px rgba(9, 105, 218, 0.5) solid !important;
   border-radius: 50% !important;
   padding: 0 !important;
   -webkit-animation: immersive-translate-loading-animation 0.6s infinite linear !important;
@@ -6418,7 +6418,7 @@ body {
   }, window.hotkeys = v2);
   var B8;
 
-  // https://esm.sh/v111/immersive-translate@1.0.8/deno/immersive-translate.js
+  // https://esm.sh/v111/immersive-translate@1.0.9/deno/immersive-translate.js
   var d4 = "Immersive Translate", v3 = class {
     #e = performance.now();
     reset() {
@@ -6717,7 +6717,7 @@ body {
       });
     }
     onFetchTokenError(e3) {
-      throw E3.error("onFetchTokenError", e3), e3 && e3.message && e3.message.indexOf("401") > -1 && this.setState({}), e3;
+      throw E3.error("onFetchTokenError", e3), e3 && e3.message && (e3.message.indexOf("401") > -1 || e3.message.indexOf("400") > -1) && this.setState({}), e3;
     }
     async updateToken() {
       let { accessTokenExpiresAt: e3 } = this.state, [t5, n3] = T6(e3);
@@ -7578,7 +7578,7 @@ body {
   for (let translation of interfaceTranslations)
     translations[translation.code] = translation.messages;
   var brandName = "Immersive Translate", brandId = "immersive-translate";
-  var brandIdForJs = "immersiveTranslate", GOOGLE_CLIENT_ID = "759003177173-mfm15s5nd77vfmo6e7lanof1emnanf0e.apps.googleusercontent.com", GOOGLE_ACCESS_TOKEN_KEY = brandIdForJs + "GoogleAccessToken", AUTH_FLOW_FLAG = brandIdForJs + "AuthFlow", LATEST_FILE_NAME = "immersive-translate-config-latest.json", AUTH_STATE_FLAG = brandIdForJs + "AuthState", iframeMessageIdentifier = brandIdForJs + "IframeMessage", iframeMessageRateIdentifier = brandIdForJs + "WaitForRateLimit", documentMessageTypeIdentifierForAsk = brandIdForJs + "DocumentMessageAsk", documentMessageTypeIdentifierForHandler = brandIdForJs + "DocumentMessageHandler", targetContainerElementAttributeName = `${brandIdForJs}Container`, specifiedTargetContainerElementAttributeName = `${brandIdForJs}SpecifiedContainer`, buildinConfigStorageKey = "buildinConfig", localConfigStorageKey = "localConfig", contextOpenOptionsMenuId = "openOptionsPage";
+  var brandIdForJs = "immersiveTranslate", GOOGLE_CLIENT_ID = "759003177173-mfm15s5nd77vfmo6e7lanof1emnanf0e.apps.googleusercontent.com", GOOGLE_ACCESS_TOKEN_KEY = brandIdForJs + "GoogleAccessToken", AUTH_FLOW_FLAG = brandIdForJs + "AuthFlow", LATEST_FILE_NAME = "immersive-translate-config-latest.json", AUTH_STATE_FLAG = brandIdForJs + "AuthState", iframeMessageIdentifier = brandIdForJs + "IframeMessage", iframeMessageRateIdentifier = brandIdForJs + "WaitForRateLimit", documentMessageTypeIdentifierForAsk = brandIdForJs + "DocumentMessageAsk", documentMessageTypeIdentifierForTellThirdParty = brandIdForJs + "DocumentMessageTellThirdParty", documentMessageTypeIdentifierForThirdPartyTell = brandIdForJs + "DocumentMessageThirdPartyTell", documentMessageTypeIdentifierForHandler = brandIdForJs + "DocumentMessageHandler", targetContainerElementAttributeName = `${brandIdForJs}Container`, specifiedTargetContainerElementAttributeName = `${brandIdForJs}SpecifiedContainer`, buildinConfigStorageKey = "buildinConfig", localConfigStorageKey = "localConfig", contextOpenOptionsMenuId = "openOptionsPage";
   var contextTranslateLocalPdfFileMenuId = "translateLocalPdfFile", contextOpenLocalEbookViewer = "openEbookViewer", contextOpenLocalEbookBuilder = "openEbookBuilder", pageTranslatedStatusEventName = `${brandIdForJs}PageTranslatedStatus`, pageUrlChangedEventName = `${brandIdForJs}PageUrlChanged`, userscriptCommandEventName = `${brandIdForJs}ReceiveCommand`, popupReceiveMessageEventName = `${brandIdForJs}PopupReceiveMessage`, hostname = "immersive-translate.owenyoung.com", homepage = `https://${hostname}/`, buildinConfigSyncUrl = `https://${hostname}/buildin_config.json`, sourceElementMarkAttributeName = `${brandIdForJs}Mark`, sourceElementEffectAttributeNameForJs = "immersiveTranslateEffect", elementMarkRootKey = `${brandIdForJs}Root`, sourceElementEffectAttributeName = `data-${brandId}-effect`, sourceElementTranslatedMarkAttributeName = `${brandIdForJs}TranslatedMark`, sourceElementParagraphAttributeName = `${brandIdForJs}ParagraphId`, sourceAtomicBlockElementMarkAttributeName = `${brandIdForJs}AtomicBlockMark`, sourceElementExcludeAttributeName = `${brandIdForJs}ExcludeMark`, sourceElementExcludeAttributeNameForSelector = `data-${brandId}-exclude-mark`, sourceElementStayOriginalAttributeName = `${brandIdForJs}StayOriginalMark`, sourcePreWhitespaceMarkAttributeName = `${brandIdForJs}PreWhitespaceMark`, sourceInlineElementMarkAttributeName = `${brandIdForJs}InlineMark`, sourceBlockElementMarkAttributeName = `${brandIdForJs}BlockMark`, sourceElementLeft = `${brandIdForJs}Left`, sourceElementRight = `${brandIdForJs}Right`, sourceElementWidth = `${brandIdForJs}Width`, sourceElementHeight = `${brandIdForJs}Height`, sourceElementTop = `${brandIdForJs}Top`, sourceElementFontSize = `${brandIdForJs}FontSize`;
   var sourceElementWithGlobalStyleMarkAttributeName = `${brandIdForJs}GlobalStyleMark`, defaultPlaceholderDelimiters = ["@", "#"], titleDelimiters = " --- ", translationTextSeparator = `
 `, translationTargetElementWrapperClass = `${brandId}-target-wrapper`, translationPdfTargetContainerClass = `${brandId}-pdf-target-container`, translationTargetInnerElementWrapperClass = `${brandId}-target-inner`, translationSourceElementsWrapperClass = `${brandId}-source-wrapper`, translationTargetTranslationElementBlockWrapperClass = `${brandId}-target-translation-block-wrapper`, translationFrameRootThemeAttributeName = `${brandId}-root-translation-theme`, translationFrameRootThemeAttributeNameForJs = `${brandIdForJs}RootTranslationTheme`, translationTargetTranslationElementVerticalBlockClass = `${brandId}-target-translation-vertical-block-wrapper`, translationTargetTranslationPdfElementBlockWrapperClass = `${brandId}-target-translation-pdf-block-wrapper`, translationTargetTranslationElementPreWhitespaceWrapperClass = `${brandId}-target-translation-pre-whitespace`, translationTargetTranslationElementInlineWrapperClass = `${brandId}-target-translation-inline-wrapper`, translationThemes = [
@@ -8591,6 +8591,15 @@ body {
   function getMainText(root2) {
     return (root2.innerText || root2.textContent || "").trim();
   }
+  function getAllIframeMainText(root2) {
+    let iframes = root2.querySelectorAll("iframe"), text = "";
+    for (let i2 = 0; i2 < iframes.length; i2++) {
+      let doc = iframes[i2].contentDocument;
+      if (doc && (text += getMainText(doc.body), text.length > 2e3))
+        break;
+    }
+    return text;
+  }
   function isMatchSelectors(selectors) {
     return selectors ? typeof selectors == "string" ? document.querySelector(selectors) !== null : selectors.some((selector) => document.querySelector(selector)) : !1;
   }
@@ -9268,7 +9277,8 @@ ${injectedCss}}
         "rp",
         "rt",
         "[spellcheck=false]",
-        ".prism-code"
+        ".prism-code",
+        "[role=code]"
       ],
       translationClasses: [],
       atomicBlockSelectors: [],
@@ -12072,7 +12082,7 @@ ${injectedCss}}
               ctx,
               currentVariableIndex
             ).currentVariableIndex, NodeFilter.FILTER_REJECT;
-          if (inlineElementGroups.length > 0) {
+          if (console.log("no inline", node2), inlineElementGroups.length > 0) {
             let paragraph = elementsToParagraph(
               [...inlineElementGroups],
               isPreWhitespaceContainer,
@@ -13201,7 +13211,8 @@ ${injectedCss}}
     }
   }
   function restorePage() {
-    clean();
+    if (clean(), !globalContext || !globalContext.mainFrame)
+      return;
     let allFrames = [
       globalContext.mainFrame,
       ...mutationObserverMap.keys()
@@ -13403,6 +13414,17 @@ ${injectedCss}}
           e3.details || ""
         );
       }), enableTitleMutationObserver(ctx));
+      let paragraphEntities2 = getParagraphEntities();
+      document.dispatchEvent(
+        new CustomEvent(documentMessageTypeIdentifierForTellThirdParty, {
+          detail: JSON.stringify({
+            type: "totalParagraphsCount",
+            payload: {
+              totalParagraphsCount: paragraphEntities2.size
+            }
+          })
+        })
+      );
     } catch (e3) {
       setPageTranslatedStatus("Error"), log_default.error(e3);
     }
@@ -13484,6 +13506,15 @@ ${injectedCss}}
       throw e3;
     }
   }
+  function setLoadingToParagraph(rootFrame, id, ctx) {
+    let element = rootFrame.querySelector(
+      "#" + translationTargetElementWrapperClass + "-" + id
+    );
+    element && (element.innerHTML = getLoadingHTML(
+      ctx.config.loadingTheme,
+      ctx.rule
+    ));
+  }
   function getLoadingHTML(theme, rule) {
     return `&#160;<${rule.targetWrapperTag} class="${brandId}-loading-${theme} notranslate"></${rule.targetWrapperTag}>`;
   }
@@ -13528,8 +13559,48 @@ ${injectedCss}}
       addParagraphToQueue(paragraph, ctx);
     setPageTranslatedStatus("Translated");
   }
+  async function retryFailedParagraphs() {
+    let allParagraphEntities = getParagraphEntities(), ids = [];
+    for (let [id, paragraph] of allParagraphEntities)
+      paragraph.state === "Error" && ids.push(id);
+    let currentParagraphLang = "auto", ctx = await getGlobalContext(getRealUrl(), {});
+    ctx.state.isDetectParagraphLanguage || (currentParagraphLang = getCurrentPageLanguage());
+    let payload = {
+      sentences: ids.filter((id) => getParagraph(id)).map((id) => {
+        let paragraph = getParagraph(id), from = paragraph.languageByLocal;
+        return from === "auto" && (from = currentParagraphLang), setLoadingToParagraph(paragraph.rootFrame, id, ctx), {
+          id: paragraph.id,
+          url: ctx.encryptedUrl,
+          text: paragraph.text,
+          from,
+          fromByClient: paragraph.languageByClient,
+          to: ctx.targetLanguage
+        };
+      })
+    };
+    if (payload.sentences.length > 0) {
+      setPageTranslatedStatus("Translating");
+      try {
+        await translateMultipleSentences(
+          payload,
+          ctx,
+          (err, translatedSentence, sentenceRequest) => {
+            onParagraphTranslated(err, translatedSentence, sentenceRequest, ctx);
+          }
+        );
+      } catch (e3) {
+        setPageTranslatedStatus("Error"), log_default.error(
+          "translateCurrentQueue error",
+          e3.name,
+          e3.message,
+          e3.details || " "
+        );
+        return;
+      }
+    }
+  }
   function onParagraphTranslated(err, translatedSentence, sentenceRequest, ctx) {
-    let paragraphWithState = getParagraph(sentenceRequest.id);
+    let translatedOk = !1, paragraphWithState = getParagraph(sentenceRequest.id);
     if (paragraphWithState && (err || !translatedSentence)) {
       err || (log_default.error("translate error", sentenceRequest, err, translatedSentence), err = new Error("no response from server"));
       let { rule } = ctx, wrapperId = sentenceRequest.id, wrapper = paragraphWithState.rootFrame.querySelector(
@@ -13565,11 +13636,21 @@ ${injectedCss}}
               sourceElementTranslatedMarkAttributeName,
               "1"
             );
-          });
+          }), translatedOk = !0;
         }
       } else
         log_default.error("paragraph not found", sentenceRequest.id);
     }
+    document.dispatchEvent(
+      new CustomEvent(documentMessageTypeIdentifierForTellThirdParty, {
+        detail: JSON.stringify({
+          type: "paragraphTranslated",
+          payload: {
+            ok: translatedOk
+          }
+        })
+      })
+    );
   }
   async function translateCurrentQueue(ctx) {
     if (currentParagraphIds.length === 0)
@@ -13720,11 +13801,20 @@ ${injectedCss}}
       meta.name = key, meta.content = ctx.rule.globalMeta[key], document.head.appendChild(meta);
     });
     let lang = ctx.sourceLanguage;
-    lang === "auto" ? (isMonkey() ? lang = await detectLanguage({
-      text: getMainText(ctx.mainFrame).slice(0, 1e3)
-    }) : isInIframe ? lang = await detectLanguage({
-      text: getMainText(ctx.mainFrame).slice(0, 1e3)
-    }) : lang = await detectTabLanguage(), lang === "auto" && (lang = await detectPageLanguage()), setCurrentPageLanguage(lang)) : setCurrentPageLanguageByClient(lang);
+    if (lang === "auto") {
+      if (!isMonkey())
+        isInIframe ? lang = await detectLanguage({
+          text: getMainText(ctx.mainFrame).slice(0, 1e3)
+        }) : lang = await detectTabLanguage();
+      else {
+        let mainText = "";
+        ctx.rule.isEbook || ctx.rule.isEbookBuilder ? mainText = getAllIframeMainText(ctx.mainFrame) : mainText = getMainText(ctx.mainFrame).slice(0, 1e3), lang = await detectLanguage({
+          text: mainText
+        });
+      }
+      lang === "auto" && (lang = await detectPageLanguage()), setCurrentPageLanguage(lang);
+    } else
+      setCurrentPageLanguageByClient(lang);
     let isAutoTranslate = ctx.state.isAutoTranslate || ctx.isTranslateUrl || ctx.rule.isPdf;
     if (!isAutoTranslate && !ctx.isTranslateExcludeUrl && (log_default.debug(`detect page language: ${lang}`), isSameTargetLanguage(lang, ctx.targetLanguage) || lang === "auto" || isMatchLanguage(lang, ctx.config.translationLanguagePattern) && (isAutoTranslate = !0, log_default.debug(`match language pattern ${lang}, auto translate`))), ctx.rule.isEbookBuilder && (isAutoTranslate = !1), isAutoTranslate)
       globalContext.state.isAutoTranslate = !0, await translatePage(globalContext);
@@ -13924,7 +14014,7 @@ ${injectedCss}}
   // content_message_listeners.ts
   var asyncMessageHandler = async function(payload, _author) {
     let { method, data } = payload;
-    method === "translateTheWholePage" ? await translateTheWholePage() : method === "translateTheMainPage" ? await translateTheMainPage() : method === "translateToThePageEndImmediately" ? await translateToThePageEndImmediately() : method === "toggleTranslatePage" ? await toggleTranslatePage() : method === "toggleTranslateTheWholePage" ? await toggleTranslateTheWholePage() : method === "toggleTranslateTheMainPage" ? await toggleTranslateTheMainPage() : method === "translatePage" ? await translatePage() : method === "toggleTranslationMask" ? await toggleTranslationMask() : method === "restorePage" ? restorePage() : method === "showTranslationOnly" ? void 0 : method === "setCurrentPageLanguageByClient" && setCurrentPageLanguageByClient(data);
+    method === "translateTheWholePage" ? await translateTheWholePage() : method === "translateTheMainPage" ? await translateTheMainPage() : method === "translateToThePageEndImmediately" ? await translateToThePageEndImmediately() : method === "toggleTranslatePage" ? await toggleTranslatePage() : method === "toggleTranslateTheWholePage" ? await toggleTranslateTheWholePage() : method === "toggleTranslateTheMainPage" ? await toggleTranslateTheMainPage() : method === "translatePage" ? await translatePage() : method === "toggleTranslationMask" ? await toggleTranslationMask() : method === "restorePage" ? restorePage() : method === "showTranslationOnly" ? void 0 : method === "setCurrentPageLanguageByClient" ? setCurrentPageLanguageByClient(data) : method === "retryFailedParagraphs" && retryFailedParagraphs();
   };
   var connection;
   function getConnection() {
@@ -15007,7 +15097,7 @@ ${injectedCss}}
     constructor() {
       super(...arguments);
       this.isSupportList = !0;
-      this.maxTextLength = 1e3;
+      this.maxTextLength = 5e3;
     }
     async translate(payload) {
       let { text } = payload, startSpaceLength = text.match(/^\s*/)[0].length;